Pull all of the query code back into connection.ex.
This structure is now essentially parallel to that in Ecto's Postgres.Connection module, which will make it far easier to track changes going forward.
