+++ lib/sqlite_ecto/error.ex
@@ -18,5 +18,4 @@ defmodule Sqlite.Ecto.Error do
   def exception({type, msg}), do: %__MODULE__{sqlite: {type, to_string(msg)}}
   def exception(msg), do: %__MODULE__{sqlite: msg}
 
-  def to_constraints(_), do: []
 end
